refactor public methods
