+++ oneflow/core/framework/tensor_impl.cpp
@@ -16,43 +16,10 @@ limitations under the License.
 #include "oneflow/core/framework/tensor_impl.h"
 #include "oneflow/core/job/parallel_desc.h"
 #include "oneflow/core/framework/device.h"
-#include "oneflow/core/control/global_process_ctx.h"
-#include "oneflow/core/common/str_util.h"
 #include "oneflow/core/framework/dtype.h"
 
 namespace oneflow {
 namespace one {
-
-namespace {
-
-std::shared_ptr<const ParallelDesc> MakeParallelDescByDevice(const Device& device) {
-  int64_t machine_id = GlobalProcessCtx::Rank() / GlobalProcessCtx::NumOfProcessPerNode();
-  int64_t device_id = device.device_id();
-  std::string machine_device_id = std::to_string(machine_id) + ":" + std::to_string(device_id);
-  ParallelConf parallel_conf;
-  parallel_conf.set_device_tag(device.of_type());
-  parallel_conf.add_device_name(machine_device_id);
-  return std::make_shared<const ParallelDesc>(parallel_conf);
-}
-
-Maybe<const Device> MakeDeviceByParallelDesc(const ParallelDesc& parallel_desc) {
-  std::string type = parallel_desc.device_tag();
-  if (parallel_desc.device_tag() == "gpu") { type = "cuda"; }
-  std::vector<std::string> machine_device_ids;
-  for (const auto& item : parallel_desc.parallel_conf().device_name()) {
-    machine_device_ids.emplace_back(item);
-  }
-  CHECK_EQ_OR_RETURN(machine_device_ids.size(), 1);
-  const std::string& machine_device_id = machine_device_ids.at(0);
-  size_t pos = machine_device_id.find(':');
-  CHECK_NE_OR_RETURN(pos, std::string::npos) << "device_name: " << machine_device_id;
-  std::string device_id = machine_device_id.substr(pos + 1);
-  CHECK_EQ_OR_RETURN(device_id.find('-'), std::string::npos);
-  CHECK_OR_RETURN(IsStrInt(device_id));
-  return std::make_shared<const Device>(type, std::stoi(device_id));
-}
-
-}  // namespace
 
 Maybe<void> TensorImpl::SyncBlobObject2Attributes(
     const std::shared_ptr<compatible_py::BlobObject>& blob_object) {
@@ -65,13 +32,13 @@ Maybe<void> TensorImpl::SyncBlobObject2Attributes(
 
 void MirroredTensorImpl::set_device(const std::shared_ptr<const Device>& device) {
   device_ = device;
-  parallel_desc_ = MakeParallelDescByDevice(*device);
+  parallel_desc_ = Device::MakeParallelDescByDevice(*device);
 }
 
 Maybe<void> MirroredTensorImpl::set_parallel_desc(
     const std::shared_ptr<const ParallelDesc>& parallel_desc) {
   parallel_desc_ = parallel_desc;
-  device_ = JUST(MakeDeviceByParallelDesc(*parallel_desc));
+  device_ = JUST(Device::MakeDeviceByParallelDesc(*parallel_desc));
   return Maybe<void>::Ok();
 }
 
@@ -79,6 +46,7 @@ Maybe<void> EagerMirroredTensorImpl::set_blob_object(
     const std::shared_ptr<compatible_py::BlobObject>& blob_object) {
   blob_object_ = blob_object;
   JUST(SyncBlobObject2Attributes(blob_object));
+  JUST(set_parallel_desc(blob_object->op_arg_parallel_attr()->parallel_desc_symbol()));
   return Maybe<void>::Ok();
 }
 
@@ -86,6 +54,7 @@ Maybe<void> EagerConsistentTensorImpl::set_blob_object(
     const std::shared_ptr<compatible_py::BlobObject>& blob_object) {
   blob_object_ = blob_object;
   JUST(SyncBlobObject2Attributes(blob_object));
+  set_parallel_desc(blob_object->op_arg_parallel_attr()->parallel_desc_symbol());
   return Maybe<void>::Ok();
 }
